fix(xpath): long long ip in dur_format's fraction path (ILP32 SegFault)
The width sweep changed the "%lld.%sS" print but left its argument
a 32-bit long — a va_arg mismatch that segfaulted the op:duration
set on the ILP32 leg only.

+++ src/leptris/xpath/evaluator_operators.c
@@ -112,7 +112,7 @@ void leptris_dur_format(double secs, char* buf, size_t cap) {
     if (rem == (double)(long)rem) {
         snprintf(sec, sizeof sec, "%lldS", (long long)rem);
     } else {
-        long ip = (long)rem;
+        long long ip = (long long)rem;
         double fr = rem - ip;
         char raw[32];
         snprintf(raw, sizeof raw, "%.9f", fr);
